Move RunOptions methods to Run
